Extract plugins into dedicated files
