+++ site/docs/.vuepress/config.ts
@@ -1,9 +1,6 @@
-import {
-  defaultTheme,
-  type DefaultThemeOptions,
-  defineUserConfig,
-} from "vuepress-vite";
-import { docsearchPlugin } from "@vuepress/plugin-docsearch";
+import { defaultTheme, defineUserConfig } from "vuepress-vite";
+import { betterLineBreaks } from "./plugins/better-line-breaks";
+import { docsearch } from "./plugins/docsearch";
 
 export default defineUserConfig({
   title: "grammY",
@@ -1063,84 +1060,19 @@ export default defineUserConfig({
     repoLabel: "GitHub",
   }),
   plugins: [
-    [
-      docsearchPlugin({
-        apiKey: "33782ffb584887e3b8cdf9e760ea8e60",
-        indexName: "grammy",
-        appId: "RBF5Q0D7QV",
-        placeholder: "Search",
-        locales: {
-          "/es/": {
-            placeholder: "Buscar",
-            translations: {
-              button: { buttonText: "Buscar" },
-            },
-          },
-          "/zh/": {
-            placeholder: "搜索文档",
-            translations: { button: { buttonText: "搜索文档" } },
-          },
-        },
-      }),
-    ],
-    [
-      {
-        name: "break-long-inline-code-snippets",
-        extendsMarkdown: (md) => {
-          md.renderer.rules.code_inline = (
-            tokens,
-            idx,
-            _opts,
-            _env,
-            slf,
-          ) => {
-            const token = tokens[idx];
-            const attributes = slf.renderAttrs(token);
-            const withBreaks = insertWbrTags(token.content);
-            const escaped = escapeHtml(withBreaks);
-            return `<code${attributes}>${escaped}</code>`;
-          };
-        },
+    docsearch({
+      "/es/": {
+        placeholder: "Buscar",
+        translations: { button: { buttonText: "Buscar" } },
       },
-    ],
+      "/zh/": {
+        placeholder: "搜索文档",
+        translations: { button: { buttonText: "搜索文档" } },
+      },
+    }),
+    betterLineBreaks(),
   ],
   markdown: {
     typographer: true,
   },
 });
-
-// Adapted from original `code_inline` implementation of markdown-it.
-const HTML_ESCAPE_TEST_RE = /&|<(?!wbr>)|(?<!<wbr)>/;
-const HTML_ESCAPE_REPLACE_RE = /&|<(?!wbr>)|(?<!<wbr)>/g;
-const HTML_REPLACEMENTS: Record<string, string> = {
-  "&": "&amp;",
-  "<": "&lt;",
-  ">": "&gt;",
-  '"': "&quot;",
-};
-function replaceUnsafeChar(ch: string) {
-  return HTML_REPLACEMENTS[ch];
-}
-function escapeHtml(str: string) {
-  return HTML_ESCAPE_TEST_RE.test(str)
-    ? str.replace(HTML_ESCAPE_REPLACE_RE, replaceUnsafeChar)
-    : str;
-}
-
-function insertWbrTags(url: string) {
-  // Adapted from https://css-tricks.com/better-line-breaks-for-long-urls/
-  return url
-    .split("//")
-    .map(
-      (str) =>
-        str
-          // Insert a word break opportunity after a colon
-          .replace(/(?<after>:)/giu, "$1<wbr>")
-          // Before a single slash, tilde, period, comma, hyphen, underline, question mark, number sign, or percent symbol
-          .replace(/(?<before>[/~.,\-_?#%])/giu, "<wbr>$1")
-          // Before and after an equals sign or ampersand
-          .replace(/(?<beforeAndAfter>[=&])/giu, "<wbr>$1<wbr>"),
-    )
-    // Reconnect the strings with word break opportunities after double slashes
-    .join("//<wbr>");
-}
